Switch ecc encrypt/decrypt to base58 encoding
